refactored rotated dataset
