+++ cordex/core/domain.py
@@ -23,11 +23,12 @@ import cf_xarray as cfxr
 import numpy as np
 import pandas as pd
 import xarray as xr
+from pyproj import CRS
 
 from ..tables import domains
 from . import cf, utils
 from .config import nround
-from .transform import transform
+from .transform import grid_mapping, transform
 
 
 def domain_names(table_name=None):
@@ -185,16 +186,13 @@ def create_dataset(
 
     x, y = _init_grid(nlon, nlat, dlon, dlat, ll_lon, ll_lat)
     if rotated is True:
-        lon, lat = rotated_coord_transform(*_stack(x, y), pollon, pollat)
-        pole = _grid_mapping(pollon, pollat)
-        if mapping_name is None:
-            mapping_name = cf.DEFAULT_MAPPING_NCVAR
         return _get_rotated_dataset(
             x,
             y,
-            lon,
-            lat,
-            pole,
+            #  lon,
+            #  lat,
+            pollon,
+            pollat,
             add_vertices=add_vertices,
             dummy=dummy,
             mapping_name=mapping_name,
@@ -303,80 +301,65 @@ def _get_regular_dataset(
 def _get_rotated_dataset(
     rlon,
     rlat,
-    lon,
-    lat,
-    pole,
+    pollon,
+    pollat,
     add_vertices=False,
     dummy=None,
     mapping_name=None,
     attrs=None,
 ):
-    data_vars = {mapping_name: pole}
+    mapping = grid_mapping(pollon, pollat, mapping_name)
+    # lon, lat = transform(x, y, src_crs=CRS.from_cf(mapping.attrs))
 
     ds = xr.Dataset(
-        data_vars=data_vars,
+        data_vars={mapping.name: mapping},
         coords=dict(
             rlon=(cf.RLON_NAME, rlon),
             rlat=([cf.RLAT_NAME], rlat),
-            lon=([cf.RLAT_NAME, cf.RLON_NAME], lon),
-            lat=([cf.RLAT_NAME, cf.RLON_NAME], lat),
         ),
         attrs=attrs,
     )
+
+    lon, lat = transform(ds.rlon, ds.rlat, src_crs=CRS.from_cf(mapping.attrs))
+    ds = ds.assign_coords(lon=lon, lat=lat)
+
+    if add_vertices is True:
+        v = vertices(ds.rlon, ds.rlat, CRS.from_cf(mapping.attrs))
+        ds = xr.merge([ds, v])
+        ds[cf.LON_NAME].attrs["bounds"] = cf.LON_BOUNDS
+        ds[cf.LAT_NAME].attrs["bounds"] = cf.LAT_BOUNDS
 
     for key, coord in ds.coords.items():
         coord.encoding["_FillValue"] = None
         coord.attrs = cf.coords[key]
 
-    if add_vertices is True:
-        from cartopy import crs as ccrs
-
-        pole = (
-            ds[mapping_name].grid_north_pole_longitude,
-            ds[mapping_name].grid_north_pole_latitude,
-        )
-        # v = vertices(ds, ccrs.RotatedPole(*pole))
-        v = vertices(ds.rlon, ds.rlat, ccrs.RotatedPole(*pole))
-        ds = xr.merge([ds, v])
-        ds[cf.LON_NAME].attrs["bounds"] = cf.LON_BOUNDS
-        ds[cf.LAT_NAME].attrs["bounds"] = cf.LAT_BOUNDS
-
-    if dummy:
-        if dummy is True:
-            dummy_name = "dummy"
-        else:
-            dummy_name = dummy
-        dummy = xr.DataArray(
-            data=np.zeros((ds.rlat.size, ds.rlon.size)),
-            coords=(ds.rlat, ds.rlon),
-        )
-        dummy.attrs = {"grid_mapping": mapping_name, "coordinates": "lat lon"}
-        ds[dummy_name] = dummy
-        if dummy_name == "topo":
-            # use cdo to create dummy topography data.
-            from cdo import Cdo
-
-            tmp = utils.get_tempfile()
-            ds.to_netcdf(tmp)
-            topo = Cdo().topo(tmp, returnXDataset=True)["topo"]
-            ds[dummy_name] = xr.DataArray(
-                data=topo.values,
-                coords=(ds.rlat, ds.rlon),
-            )
-            ds[dummy_name].attrs.update(topo.attrs)
+    if dummy is not None:
+        ds = _add_dummy(ds, dummy, mapping)
     return ds
 
 
-def _grid_mapping(pollon, pollat, mapping_name=None):
-    """creates a grid mapping DataArray object"""
-    if mapping_name is None:
-        mapping_name = cf.DEFAULT_MAPPING_NCVAR
-    da = xr.DataArray(np.zeros((), dtype=np.int32))
-    attrs = cf.mapping.copy()
-    attrs["grid_north_pole_longitude"] = pollon
-    attrs["grid_north_pole_latitude"] = pollat
-    da.attrs = attrs
-    return da
+def _add_dummy(ds, name=True, mapping=None):
+    if name is True:
+        name = "dummy"
+    dummy = xr.DataArray(
+        data=np.zeros((ds.rlat.size, ds.rlon.size)),
+        coords=(ds.rlat, ds.rlon),
+    )
+    dummy.attrs = {"grid_mapping": mapping.name, "coordinates": "lat lon"}
+    ds[name] = dummy
+    if name == "topo":
+        # use cdo to create dummy topography data.
+        from cdo import Cdo
+
+        tmp = utils.get_tempfile()
+        ds.to_netcdf(tmp)
+        topo = Cdo().topo(tmp, returnXDataset=True)["topo"]
+        ds[name] = xr.DataArray(
+            data=topo.values,
+            coords=(ds.rlat, ds.rlon),
+        )
+        ds[name].attrs.update(topo.attrs)
+    return ds
 
 
 def _init_grid(nlon, nlat, dlon, dlat, ll_lon, ll_lat):
@@ -397,92 +380,6 @@ def _stack(x, y):
     return np.vstack(len(tmp_y) * (tmp_x,)), np.hstack(
         len(tmp_x) * (tmp_y[:, np.newaxis],)
     )
-
-
-def rotated_coord_transform(lon, lat, np_lon, np_lat, direction="rot2geo"):
-    """Transforms a coordinate into a rotated grid coordinate and vice versa.
-
-    The coordinates have to be given in degree and will be returned in degree.
-
-    Parameters
-    ----------
-    lon : float array like
-        Longitude coordinate.
-    lat : float array like
-        Latitude coordinate.
-    np_lon : float array like
-        Longitude coordinate of the rotated north pole.
-    np_lat : float array like
-        Latitude coordinate of the rotated north pole.
-    direction : str
-        Direction of the rotation.
-        Options are: 'rot2geo' (default) for a transformation to regular
-        coordinates from rotated. 'geo2rot' transforms regular coordinates
-        to rotated.
-
-    Returns
-    -------
-    lon_new : array like
-        New longitude coordinate.
-    lat_new : array like
-        New latitude coordinate.
-    """
-    # Convert degrees to radians
-    lon = np.deg2rad(lon)
-    lat = np.deg2rad(lat)
-
-    theta = 90.0 - np_lat  # Rotation around y-axis
-    phi = np_lon + 180.0  # Rotation around z-axis
-
-    # Convert degrees to radians
-    phi = np.deg2rad(phi)
-    theta = np.deg2rad(theta)
-
-    # Convert from spherical to cartesian coordinates
-    x = np.cos(lon) * np.cos(lat)
-    y = np.sin(lon) * np.cos(lat)
-    z = np.sin(lat)
-
-    # Regular -> Rotated
-    if direction == "geo2rot":
-        x_new = (
-            np.cos(theta) * np.cos(phi) * x
-            + np.cos(theta) * np.sin(phi) * y
-            + np.sin(theta) * z
-        )
-        y_new = -np.sin(phi) * x + np.cos(phi) * y
-        z_new = (
-            -np.sin(theta) * np.cos(phi) * x
-            - np.sin(theta) * np.sin(phi) * y
-            + np.cos(theta) * z
-        )
-
-    # Rotated -> Regular
-    elif direction == "rot2geo":
-        phi = -phi
-        theta = -theta
-
-        x_new = (
-            np.cos(theta) * np.cos(phi) * x
-            + np.sin(phi) * y
-            + np.sin(theta) * np.cos(phi) * z
-        )
-        y_new = (
-            -np.cos(theta) * np.sin(phi) * x
-            + np.cos(phi) * y
-            - np.sin(theta) * np.sin(phi) * z
-        )
-        z_new = -np.sin(theta) * x + np.cos(theta) * z
-
-    # Convert cartesian back to spherical coordinates
-    lon_new = np.arctan2(y_new, x_new)
-    lat_new = np.arcsin(z_new)
-
-    # Convert radians back to degrees
-    lon_new = np.rad2deg(lon_new)
-    lat_new = np.rad2deg(lat_new)
-
-    return lon_new, lat_new
 
 
 def _dcoord(coord, include="left"):
@@ -628,54 +525,3 @@ def vertices(rlon, rlat, src_crs, trg_crs=None):
     lon_vertices.attrs = cf.coords[cf.LON_BOUNDS]
     lat_vertices.attrs = cf.coords[cf.LAT_BOUNDS]
     return xr.merge([lat_vertices, lon_vertices])
-
-
-def lon_lat_bounds_coordinates(ds, src_crs=None, trg_crs=None):
-    """Compute lon and lat bounds coordinates.
-
-    Transformation of rlon vertices and rlat vertices
-    into the target crs according to
-    https://cfconventions.org/cf-conventions/cf-conventions.html#cell-boundaries
-
-    Parameters
-    ----------
-    rlon : xr.DataArray
-        Longitude in rotated pole grid.
-    rlat : xr.DataArray
-        Latitude in rotated pole grid.
-
-    Returns
-    -------
-    vertices : xr.Dataset
-        lon_vertices and lat_vertices in target crs.
-
-    """
-    if src_crs is None:
-        src_crs = get_pole_crs(ds)
-    rot_bounds = bounds_coordinates(ds, ("rlon", "rlat"))
-    lon_bounds, lat_bounds = transform(
-        rot_bounds.coords["rlon_b"], rot_bounds.coords["rlat_b"], src_crs, trg_crs
-    )
-    return ds.assign_coords(lon_b=lon_bounds.transpose(), lat_b=lat_bounds.transpose())
-
-
-def get_pole_crs(ds, grid_mapping_name="rotated_latitude_longitude"):
-    """Create cartopy crs from rotated pole dataset."""
-    from cartopy import crs as ccrs
-
-    pole = (
-        ds[grid_mapping_name].grid_north_pole_longitude,
-        ds[grid_mapping_name].grid_north_pole_latitude,
-    )
-    return ccrs.RotatedPole(*pole)
-
-
-def extend_coordinate(coord, n=1):
-    data = coord
-    # insert = np.array([data[0] - n*(data[1]-data[0]) + i*(data[1]-data[0]) for i in range(0, n)])
-    insert = np.array([data[0] - (n - i) * (data[1] - data[0]) for i in range(0, n)])
-    return insert
-    data_extended = np.insert(data, 0, data[0] - (data[1] - data[0]))
-    # return data_extended
-    # data_extended = np.append(data_extended, data[-1]+(data[-1]-data[-2]))
-    return xr.DataArray(data_extended, dims=coord.dims, name=coord.name)
